test: pass basic function test

+++ pkg/serverless/function/image.go
@@ -5,6 +5,7 @@ import (
 	"io"
 	"os"
 	"os/exec"
+	"strings"
 )
 
 const serverIp = "localhost"
@@ -75,15 +76,52 @@ func SaveImage(name string) error {
 	return nil
 }
 
+
+// find the image
+func FindImage(name string) bool {
+	cmd := exec.Command("docker", "images", name)
+
+	// check the output
+	output, err := cmd.CombinedOutput()
+	if err != nil {
+		log.Error("[FindImage] get output error: ", err)
+		return false
+	}
+
+	result := strings.TrimSpace(string(output))
+	log.Info("[FindImage] the result is: ", result)
+
+	if strings.Contains(result, name) {
+		return true
+	} else {
+		return false
+	}
+}
+
+
 // DeleteImage to delete image for function
 func DeleteImage(name string) error {
+	// if the image not exist, just ignore
 	imageName := serverIp + ":5000/" + name + ":latest"
-	cmd := exec.Command("docker", "rmi", imageName)
-	err := cmd.Run()
-	if err != nil {
-		log.Error("[DeleteImage] delete image error: ", err)
-		return err
+	if FindImage(imageName) {
+		cmd := exec.Command("docker", "rmi", imageName)
+		err := cmd.Run()
+		if err != nil {
+			log.Error("[DeleteImage] delete first image error: ", err)
+			return err
+		}
 	}
+	
+	if FindImage(name) {
+		cmd := exec.Command("docker", "rmi", name + ":latest") 
+		err := cmd.Run()
+		if err != nil {
+			log.Error("[DeleteImage] delete second image error: ", err)
+			return err
+		}
+	}
+	
+	log.Info("[DeleteImage] delete image finished")
 	return nil
 }
